backend/requirements.txt
fastapi[all]
uvicorn
python-multipart
pymupdf
sentence-transformers
faiss-cpu
numpy
